pyproject.toml
[project]
name = "repl-nix-carpicsorterai-1"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.2.0",
    "flask-login>=0.6.3",
    "numpy>=1.26.4",
    "pandas>=2.2.3",
    "pillow>=10.4.0",
    "plotly>=5.24.1",
    "psycopg2-binary>=2.9.10",
    "streamlit>=1.39.0",
    "tensorflow-cpu>=2.17.0",
]
